Modify the totalchi oracle to avoid the totalchi display is not normal

+++ app/src/main/java/io/pp/net_disk_demo/activity/RenewActivity.java
@@ -335,7 +335,7 @@ public class RenewActivity extends BaseActivity implements RenewView {
     }
 
     @Override
-    public void showGetTotalChiView(int totalChi) {
+    public void showGetTotalChiView(long totalChi) {
         runOnUiThread(new Runnable() {
             @Override
             public void run() {
