+++ app/src/main/java/io/pp/net_disk_demo/mvp/presenter/ProphecyPresenter.java
@@ -12,7 +12,7 @@ public interface ProphecyPresenter {
 
     void showRequestTotalChi();
 
-    void showGetTotalChi(int totalChi);
+    void showGetTotalChi(long totalChi);
 
     void showGetTotalChiFailed(String errMsg);
