+++ app/src/main/java/io/pp/net_disk_demo/mvp/presenter/UploadPresenter.java
@@ -58,7 +58,7 @@ public interface UploadPresenter {
 
     void showRequestTotalChi();
 
-    void showGetTotalChi(int totalChi);
+    void showGetTotalChi(long totalChi);
 
     void showGetTotalChiFailed(String errMsg);
