+++ app/src/main/java/io/pp/net_disk_demo/mvp/presenter/presenterimpl/ProphecyPresenterImpl.java
@@ -50,7 +50,7 @@ public class ProphecyPresenterImpl implements ProphecyPresenter {
     }
 
     @Override
-    public void showGetTotalChi(int totalChi) {
+    public void showGetTotalChi(long totalChi) {
         if (mProphecyView != null) {
             mProphecyView.showGetTotalChiView(totalChi);
         }
