+++ app/src/main/java/io/pp/net_disk_demo/mvp/view/ProphecyView.java
@@ -4,7 +4,7 @@ public interface ProphecyView {
 
     void showRequestTotalChiView();
 
-    void showGetTotalChiView(int totalChi);
+    void showGetTotalChiView(long totalChi);
 
     void showGetTotalChiFailedView(String errMsg);
