+++ app/src/main/java/io/pp/net_disk_demo/mvp/view/RenewView.java
@@ -33,7 +33,7 @@ public interface RenewView {
 
     void showRequestTotalChiView();
 
-    void showGetTotalChiView(int totalChi);
+    void showGetTotalChiView(long totalChi);
 
     void showGetTotalChiFailedView(String errMsg);
 }
